requirements.txt
aiohttp
fastapi==0.110.3
h11
httpcore
httpx
numpy
openai
packaging
passlib
pydantic
pydantic-settings
pydantic_core
#PyExecJS==1.5.1
PyJWT
PyMySQL
python-dotenv
python-multipart
starlette
sse-starlette
typing_extensions
urllib3
uvicorn
#websockets
python-pptx
tiktoken
SQLAlchemy
pandas
pypandoc
python-docx
#markdown_to_json
asgiref
spacy
weaviate-client
pdfplumber
pypdf
#watchdog